AWSMobileClient moved the user attritbutes to separate class
