+++ AWSAuthSDK/Sources/AWSMobileClient/AWSMobileClientExtensions.swift
@@ -362,7 +362,7 @@ extension AWSMobileClient {
                 let confirmedStatus: SignUpConfirmationState = .unconfirmed
                 var codeDeliveryDetails: UserCodeDeliveryDetails? = nil
                 if let deliveryDetails = result.codeDeliveryDetails {
-                    codeDeliveryDetails = self.getUserCodeDeliveryDetails(deliveryDetails)
+                    codeDeliveryDetails = UserCodeDeliveryDetails.getUserCodeDeliveryDetails(deliveryDetails)
                 }
                 completionHandler(SignUpResult(signUpState: confirmedStatus, codeDeliveryDetails: codeDeliveryDetails), nil)
             }
@@ -383,7 +383,7 @@ extension AWSMobileClient {
             } else if let result = task.result {
                 var codeDeliveryDetails: UserCodeDeliveryDetails? = nil
                 if let deliveryDetails = result.codeDeliveryDetails {
-                    codeDeliveryDetails = self.getUserCodeDeliveryDetails(deliveryDetails)
+                    codeDeliveryDetails = UserCodeDeliveryDetails.getUserCodeDeliveryDetails(deliveryDetails)
                 }
                 completionHandler(ForgotPasswordResult(forgotPasswordState: .confirmationCodeSent, codeDeliveryDetails: codeDeliveryDetails), nil)
             }
@@ -558,111 +558,6 @@ extension AWSMobileClient {
             return nil
         }
     }
-
-    /// Verify a user attribute like phone_number.
-    ///
-    /// - Parameters:
-    ///   - attributeName: name of the attribute.
-    ///   - completionHandler: completionHandler which will be called when the result is avilable.
-    public func verifyUserAttribute(attributeName: String, completionHandler: @escaping ((UserCodeDeliveryDetails?, Error?) -> Void)) {
-        self.userpoolOpsHelper.currentActiveUser!.getAttributeVerificationCode(attributeName).continueWith { (task) -> Any? in
-            if let error = task.error {
-                completionHandler(nil, AWSMobileClientError.makeMobileClientError(from: error))
-            } else if let result = task.result {
-                var codeDeliveryDetails: UserCodeDeliveryDetails? = nil
-                if let deliveryDetails = result.codeDeliveryDetails {
-                    codeDeliveryDetails = self.getUserCodeDeliveryDetails(deliveryDetails)
-                }
-                completionHandler(codeDeliveryDetails, nil)
-            }
-            return nil
-        }
-    }
-
-    /// Update the attributes for a user.
-    ///
-    /// - Parameters:
-    ///   - attributeMap: the attribute map of the user.
-    ///   - completionHandler: completionHandler which will be called when the result is avilable.
-    public func updateUserAttributes(attributeMap: [String: String], completionHandler: @escaping (([UserCodeDeliveryDetails]?, Error?) -> Void)) {
-        let attributes = attributeMap.map {AWSCognitoIdentityUserAttributeType.init(name: $0, value: $1) }
-        self.userpoolOpsHelper.currentActiveUser!.update(attributes).continueWith { (task) -> Any? in
-            if let error = task.error {
-                completionHandler(nil, AWSMobileClientError.makeMobileClientError(from: error))
-            } else if let result = task.result {
-                var codeDeliveryDetailsList: [UserCodeDeliveryDetails] = []
-                if result.codeDeliveryDetailsList != nil {
-                    for codeDeliveryDetail in result.codeDeliveryDetailsList! {
-                        codeDeliveryDetailsList.append(self.getUserCodeDeliveryDetails(codeDeliveryDetail))
-                    }
-                }
-                completionHandler(codeDeliveryDetailsList, nil)
-            }
-            return nil
-        }
-    }
-
-    /// Fetches the attributes for logged in user.
-    ///
-    /// - Parameter completionHandler: completion handler which will be invoked when result is available.
-    public func getUserAttributes(completionHandler: @escaping (([String: String]?, Error?) -> Void)) {
-        self.userpoolOpsHelper.currentActiveUser!.getDetails().continueWith { (task) -> Any? in
-            if let error = task.error {
-               completionHandler(nil, AWSMobileClientError.makeMobileClientError(from: error))
-            } else if let result = task.result {
-                let userAttributes = result.userAttributes
-                var attributesMap = [String: String]()
-                if let userAttributes = userAttributes {
-                    for attribute in userAttributes {
-                        guard attribute.name != nil, attribute.value != nil else {continue}
-                        attributesMap[attribute.name!] = attribute.value!
-                    }
-                }
-                completionHandler(attributesMap, nil)
-            }
-            return nil
-        }
-    }
-
-    /// Confirm the updated attributes using a confirmation code.
-    ///
-    /// - Parameters:
-    ///   - attributeName: the attribute to be confirmed.
-    ///   - code: the code sent to the user.
-    ///   - completionHandler: completionHandler which will be called when the result is avilable.
-    public func confirmUpdateUserAttributes(attributeName: String, code: String, completionHandler: @escaping ((Error?) -> Void)) {
-        self.confirmVerifyUserAttribute(attributeName: attributeName, code: code, completionHandler: completionHandler)
-    }
-
-    /// Confirm the attribute using a confirmation code.
-    ///
-    /// - Parameters:
-    ///   - attributeName: the attribute to be verified.
-    ///   - code: the code sent to the user.
-    ///   - completionHandler: completionHandler which will be called when the result is avilable.
-    public func confirmVerifyUserAttribute(attributeName: String, code: String, completionHandler: @escaping ((Error?) -> Void)) {
-        self.userpoolOpsHelper.currentActiveUser!.verifyAttribute(attributeName, code: code).continueWith { (task) -> Any? in
-            if let error = task.error {
-                completionHandler(AWSMobileClientError.makeMobileClientError(from: error))
-            } else if let _ = task.result {
-                completionHandler(nil)
-            }
-            return nil
-        }
-    }
-
-    internal func getUserCodeDeliveryDetails(_ deliveryDetails: AWSCognitoIdentityProviderCodeDeliveryDetailsType) -> UserCodeDeliveryDetails {
-        var codeDeliveryDetails: UserCodeDeliveryDetails?
-        switch(deliveryDetails.deliveryMedium) {
-        case .email:
-            codeDeliveryDetails = UserCodeDeliveryDetails(deliveryMedium: .email, destination: deliveryDetails.destination, attributeName: deliveryDetails.attributeName)
-        case .sms:
-            codeDeliveryDetails = UserCodeDeliveryDetails(deliveryMedium: .sms, destination: deliveryDetails.destination, attributeName: deliveryDetails.attributeName)
-        case .unknown:
-            codeDeliveryDetails = UserCodeDeliveryDetails(deliveryMedium: .unknown, destination: deliveryDetails.destination, attributeName: deliveryDetails.attributeName)
-        }
-        return codeDeliveryDetails!
-    }
  
     /// Fetches the `AWSCredentials` asynchronously.
     ///
@@ -809,6 +704,64 @@ extension AWSMobileClient {
 
 }
 
+//MARK: Extension to hold user attribute operations
+extension AWSMobileClient {
+    
+    /// Verify a user attribute like phone_number.
+    ///
+    /// - Parameters:
+    ///   - attributeName: name of the attribute.
+    ///   - completionHandler: completionHandler which will be called when the result is avilable.
+    public func verifyUserAttribute(attributeName: String,
+                                    completionHandler: @escaping ((UserCodeDeliveryDetails?, Error?) -> Void)) {
+        let userDetails = AWSMobileClientUserDetails(withUser: self.userpoolOpsHelper.currentActiveUser!)
+        userDetails.verifyUserAttribute(attributeName: attributeName, completionHandler: completionHandler)
+    }
+    
+    /// Update the attributes for a user.
+    ///
+    /// - Parameters:
+    ///   - attributeMap: the attribute map of the user.
+    ///   - completionHandler: completionHandler which will be called when the result is avilable.
+    public func updateUserAttributes(attributeMap: [String: String],
+                                     completionHandler: @escaping (([UserCodeDeliveryDetails]?, Error?) -> Void)) {
+        let userDetails = AWSMobileClientUserDetails(withUser: self.userpoolOpsHelper.currentActiveUser!)
+        userDetails.updateUserAttributes(attributeMap: attributeMap, completionHandler: completionHandler)
+    }
+    
+    /// Fetches the attributes for logged in user.
+    ///
+    /// - Parameter completionHandler: completion handler which will be invoked when result is available.
+    public func getUserAttributes(completionHandler: @escaping (([String: String]?, Error?) -> Void)) {
+        let userDetails = AWSMobileClientUserDetails(withUser: self.userpoolOpsHelper.currentActiveUser!)
+        userDetails.getUserAttributes(completionHandler: completionHandler)
+    }
+    
+    /// Confirm the updated attributes using a confirmation code.
+    ///
+    /// - Parameters:
+    ///   - attributeName: the attribute to be confirmed.
+    ///   - code: the code sent to the user.
+    ///   - completionHandler: completionHandler which will be called when the result is avilable.
+    public func confirmUpdateUserAttributes(attributeName: String, code: String, completionHandler: @escaping ((Error?) -> Void)) {
+        self.confirmVerifyUserAttribute(attributeName: attributeName, code: code, completionHandler: completionHandler)
+    }
+    
+    /// Confirm the attribute using a confirmation code.
+    ///
+    /// - Parameters:
+    ///   - attributeName: the attribute to be verified.
+    ///   - code: the code sent to the user.
+    ///   - completionHandler: completionHandler which will be called when the result is avilable.
+    public func confirmVerifyUserAttribute(attributeName: String, code: String, completionHandler: @escaping ((Error?) -> Void)) {
+        let userDetails = AWSMobileClientUserDetails(withUser: self.userpoolOpsHelper.currentActiveUser!)
+        userDetails.confirmVerifyUserAttribute(attributeName: attributeName,
+                                               code: code,
+                                               completionHandler: completionHandler)
+    }
+}
+
+//MARK: Extension to hold UserPoolAuthHelperlCallbacks methods
 extension AWSMobileClient: UserPoolAuthHelperlCallbacks {
     
     func getPasswordDetails(_ authenticationInput: AWSCognitoIdentityPasswordAuthenticationInput, passwordAuthenticationCompletionSource: AWSTaskCompletionSource<AWSCognitoIdentityPasswordAuthenticationDetails>) {
